opt: disallow non-standard NULL ordering with subqueries in test builds
The optimizer does not correctly build `ORDER BY` expressions with
non-standard `NULL` orderings when one or more of the ordering columns
is a projection containing a subquery. This causes test-only assertions
to fail in randomized tests. It is not yet known if there are any issues
with the results of these queries in production builds where the test
assertions are disabled.

For now, we simply return an "unimplemented" error to silence the
failures.

Informs #129956

Release note: None